diocane

+++ project/include/train.hpp
@@ -14,11 +14,11 @@ protected:
     double distance;
     const int direction;
     int late = 0;
-    std::vector<int> expected_times;
+    std::vector<double> expected_times;
     int stations_done = 0;
 
 public:
-    Train(int i, int d, int n, std::vector<int> t);
+    Train(int i, int d, int n, std::vector<double> t);
     //~Train();
 
     inline int    get_train_name()       const { return train_name       ; }
@@ -35,19 +35,19 @@ public:
 class Fast_Train : public Train
 {
 public:
-    Fast_Train(int d, int n, std::vector<int> t) : Train(300, d, n, t) {}
+    Fast_Train(int d, int n, std::vector<double> t) : Train(300, d, n, t) {}
 };
 
 class Medium_Train : public Train
 {
 public:
-    Medium_Train(int d, int n, std::vector<int> t) : Train(240, d, n, t) {}
+    Medium_Train(int d, int n, std::vector<double> t) : Train(240, d, n, t) {}
 };
 
 class Slow_Train : public Train
 {
 public:
-    Slow_Train(int d, int n, std::vector<int> t) : Train(160, d, n, t) {}
+    Slow_Train(int d, int n, std::vector<double> t) : Train(160, d, n, t) {}
 };
 
 #endif
